Simon - Merge vote élection et candidature crud handle

+++ vote/views1/candidate.py
@@ -1,5 +1,3 @@
-
-
 from vote.models import Candidate
 from vote.serializers import CandidateSerializer
 from rest_framework.views import APIView
@@ -7,19 +5,38 @@ from rest_framework import response, status
 from django.http import Http404
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
-from vote.views1.user import res
+from vote.serializers.candidate import CandidateApprouveSerializer
+from vote.views1.user import CustomAuthentication, res
+from vote.paginations import CustomPaginator
+from rest_framework.pagination import PageNumberPagination
+from rest_framework.permissions import IsAdminUser
 
 
 class CandidateView(APIView):
+    authentication_classes = [CustomAuthentication]
 
     @swagger_auto_schema(
         operation_description="Returns candidature list",
         responses= res
     )
     def get(self, request):
-        candidate = candidate.objects.all()
-        serializer = CandidateSerializer(candidate, many=True)
-        return response.Response(serializer.validated_data, status=status.HTTP_200_OK)
+
+        if request.user.is_authenticated and  request.user.has_perm('vote.view_candidate'):
+            candidates = Candidate.objects.all()
+            paginator =PageNumberPagination()
+            paginator_queryset = paginator.paginate_queryset(candidates, request)
+            serializer = CandidateSerializer(paginator_queryset, many=True)
+            # print("results", paginator.get_results(serializer.data))
+            return response.Response({
+                "data": CustomPaginator.format_json_response(paginator, serializer.data), # , serializer.data
+                "details": "Liste des candidatures",
+                "succes": True
+            }, status=status.HTTP_200_OK)
+            
+        return response.Response({
+            "details": "Access denied",
+            "succes": False
+        }, status=status.HTTP_403_FORBIDDEN)
 
     @swagger_auto_schema(
         operation_description="Create new candidature",
@@ -27,13 +44,20 @@ class CandidateView(APIView):
         responses= res
     )
     def post(self, request):
-        serializer = CandidateSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return response.Response(serializer.validated_data, status=status.HTTP_201_CREATED)
-        return response.Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated and  request.user.has_perm('vote.add_candidate'):
+            serializer = CandidateSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return response.Response(serializer.data, status=status.HTTP_201_CREATED)
+            return response.Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return response.Response({
+            "details": "Access denied",
+            "succes": False
+        }, status=status.HTTP_403_FORBIDDEN)
 
 class CandidateDetailView(APIView):
+    authentication_classes = [CustomAuthentication]
+    
     def get_object(self, pk):
         try:
             return Candidate.objects.get(pk=pk)
@@ -45,14 +69,20 @@ class CandidateDetailView(APIView):
         responses= res
     ) 
     def get(self, request, pk):
-        candidate = self.get_object(pk)
-        if not candidate:
-            return response.Response(
-                {"error": "Candidature non trouvée"},
-                status=status.HTTP_404_NOT_FOUND
-            )
-        serializer = CandidateSerializer(candidate)
-        return response.Response(serializer.validated_data, status=status.HTTP_200_OK)
+        if request.user.is_authenticated and  request.user.has_perm('vote.view_candidate'):
+            candidate = self.get_object(pk)
+            if not candidate:
+                return response.Response(
+                    {"error": "Candidature non trouvée"},
+                    status=status.HTTP_404_NOT_FOUND
+                )
+            serializer = CandidateSerializer(candidate)
+            return response.Response(serializer.data, status=status.HTTP_200_OK)
+        return response.Response({
+            "details": "Access denied",
+            "succes": False
+        }, status=status.HTTP_403_FORBIDDEN)
+
 
     @swagger_auto_schema(
         operation_description="Modify a single candidature details",
@@ -60,17 +90,22 @@ class CandidateDetailView(APIView):
         responses= res
     )    
     def put(self, request, pk):
-        candidate = self.get_object(pk)
-        if not candidate:
-            return response.Response(
-                {"error": "Candidature non trouvée"},
-                status=status.HTTP_404_NOT_FOUND
-            )
-        serializer = CandidateSerializer(candidate, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return response.Response(serializer.validated_data, status=status.HTTP_200_OK)
-        return response.Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated and  request.user.has_perm('vote.change_candidate'):
+            candidate = self.get_object(pk)
+            if not candidate:
+                return response.Response(
+                    {"error": "Candidature non trouvée"},
+                    status=status.HTTP_404_NOT_FOUND
+                )
+            serializer = CandidateSerializer(candidate, data=request.data, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return response.Response(serializer.data, status=status.HTTP_200_OK)
+            return response.Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return response.Response({
+            "details": "Access denied",
+            "succes": False
+        }, status=status.HTTP_403_FORBIDDEN)
     
     @swagger_auto_schema(
         operation_description="Delete a single candidature by id",
@@ -81,11 +116,44 @@ class CandidateDetailView(APIView):
         }
     )
     def delete(self, request, pk):
-        candidate= self.get_object(pk)
-        if not candidate:
-            return response.Response(
-                {"error": "Candidature non trouvée"},
-                status=status.HTTP_404_NOT_FOUND
-            )
-        candidate.delete()
-        return response.Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.is_authenticated and  request.user.has_perm('vote.delete_candidate'):
+            candidate= self.get_object(pk)
+            if not candidate:
+                return response.Response(
+                    {"error": "Candidature non trouvée"},
+                    status=status.HTTP_404_NOT_FOUND
+                )
+            candidate.delete()
+            return response.Response(status=status.HTTP_204_NO_CONTENT)
+        return response.Response({
+            "details": "Access denied",
+            "succes": False
+        }, status=status.HTTP_403_FORBIDDEN)
+
+class CandidateApprouveView(APIView):
+    authentication_classes = [CustomAuthentication]
+    permission_classes = [IsAdminUser] # only for supervisor or admin
+    
+    @swagger_auto_schema(
+        operation_description="Modify a single candidature details",
+        request_body=CandidateApprouveSerializer,
+        responses= res
+    )    
+    def put(self, request, pk):
+        if request.user.is_authenticated and  request.user.has_perm('vote.change_candidate'):
+            candidate = self.get_object(pk)
+            if not candidate:
+                return response.Response(
+                    {"error": "Candidature non trouvée"},
+                    status=status.HTTP_404_NOT_FOUND
+                )
+            serializer = CandidateApprouveSerializer(candidate, data=request.data, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return response.Response(serializer.data, status=status.HTTP_200_OK)
+            return response.Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return response.Response({
+            "details": "Access denied",
+            "succes": False
+        }, status=status.HTTP_403_FORBIDDEN)
+
